Create README.md
Play With Sort(PWS): It's an web application that is useful to understand the working of Sorting algorithms with ease even to a fresher. It has visual bars and some sorting buttons which when selected it will automatically sort the random bars generated.It is developed with a goal of making sorting fun and at the same time feel comfort to without memorizing it with difficult.